requirements.txt
duckdb
fastapi
uvicorn[standard]
apache-airflow-providers-amazon
s3fs
boto3
python-dotenv
SQLAlchemy